mcp-mssql-server/pyproject.toml
[project]
name = "mcp-mssql-server"
version = "0.1.0"
description = "A Model Context Protocol server for Microsoft SQL Server with data analysis and visualization capabilities"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"  # Changed from {text = "MIT"} to simple string
authors = [
    {name = "Serena XiaoNaHang Sinclair", email = "serena_hang@hotmail.com"}
]
keywords = ["mcp", "mssql", "sql-server", "database", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.28.1",
    "fastmcp>=0.1.0",
    "pymssql>=2.2.8",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    "nbformat>=5.9.0",
    "ipykernel>=6.25.0",
    "neo4j>=5.14.0",
    "networkx>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/SerenaHangSinclair/mcp-mssql-server"
Repository = "https://github.com/SerenaHangSinclair/mcp-mssql-server"


[project.scripts]
mcp-mssql-server = "mssql:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["mssql", "main"]



